Move MainModelAbstract to its own page
